测试22，23更新
1. AE部分

1.1 单位上没有新AE的时候不会使Ares的隐形反隐AE失效

1.2 修正了给死亡单位附加AE或结束动画导致FE的bug

1.3 增加随机选取动画和结束动画的条目

1.4 增加覆盖InfDeath的条目

2. 修正乘客移动到死亡状态子机时FE的bug

3. 修正自定义血条不读取全局设置的bug

4. 烈葱方向护甲添加

5.(测试21YRpp) 修正虚函数参数错误导致随机FE的bug

+++ src/Ext/WarheadType/Detonate.cpp
@@ -438,7 +438,7 @@ void WarheadTypeExt::ExtData::ApplyShieldModifiers(TechnoClass* pTarget)
 			if (this->Shield_ReplaceOnly)
 			{
 				if (shieldIndex >= 0)
-					shieldType = Shield_AttachTypes[Math::min(shieldIndex, (signed)Shield_AttachTypes.size() - 1)];
+					shieldType = Shield_AttachTypes[Math::min(shieldIndex, static_cast<int>(Shield_AttachTypes.size()) - 1)];
 			}
 			else
 			{
@@ -1373,10 +1373,10 @@ void WarheadTypeExt::ExtData::ApplyDirectional(BulletClass* pBullet)
 	if (!pTarTypeExt->DirectionalArmor || pTarget->WhatAmI() != AbstractType::Unit || pBullet->Type->Vertical)
 		return;
 
-	pTarTypeExt->DirectionalArmor_FrontField = Math::min(pTarTypeExt->DirectionalArmor_FrontField, 1.0);
-	pTarTypeExt->DirectionalArmor_FrontField = Math::max(pTarTypeExt->DirectionalArmor_FrontField, 0);
-	pTarTypeExt->DirectionalArmor_BackField = Math::min(pTarTypeExt->DirectionalArmor_BackField, 1.0);
-	pTarTypeExt->DirectionalArmor_BackField = Math::max(pTarTypeExt->DirectionalArmor_BackField, 0);
+	pTarTypeExt->DirectionalArmor_FrontField = Math::min(pTarTypeExt->DirectionalArmor_FrontField, 1.0f);
+	pTarTypeExt->DirectionalArmor_FrontField = Math::max(pTarTypeExt->DirectionalArmor_FrontField, 0.0f);
+	pTarTypeExt->DirectionalArmor_BackField = Math::min(pTarTypeExt->DirectionalArmor_BackField, 1.0f);
+	pTarTypeExt->DirectionalArmor_BackField = Math::max(pTarTypeExt->DirectionalArmor_BackField, 0.0f);
 
 	const int tarFacing = pTarget->PrimaryFacing.current().value256();
 	int bulletFacing = BulletExt::ExtMap.Find(pBullet)->BulletDir.value256();
